@@ -6,8 +6,9 @@
   "dependencies": {
     "bootstrap": "^3.3.7",
     "ejs": "~2.5.6",
-    "react": "^15.4.2",
-    "react-bootstrap": "^0.30.10",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
+    "react-bootstrap": "^0.31.1",
     "react-dom": "^15.4.2",
     "react-router": "^3.0.3",
     "react-router-bootstrap": "^0.23.3"
