@@ -2,7 +2,8 @@
  * @author - Greg Wolverson
  */
 import React, {Component} from 'react';
-import {Button, Col, Panel, Row, Table} from 'react-bootstrap';
+import PropTypes from 'prop-types';
+import {Alert, Button, Col, Glyphicon, Panel, Row, Table} from 'react-bootstrap';
 import {LinkContainer} from 'react-router-bootstrap';
 
 class DataRow extends Component {
@@ -14,23 +15,25 @@ class DataRow extends Component {
                 </LinkContainer>
                 <td>{this.props.teamMember.resourceName}</td>
                 <td>{this.props.teamMember.role}</td>
-                <LinkContainer to={'/team/' + this.props.team.teamName
-                + '/' + this.props.teamMember._id + '/delete'}>
-                    <Button href bsStyle="danger" block>Remove</Button>
-                </LinkContainer>
+                <a className="btn btn-danger" onClick={(event) => this.context.removeResource(this.props.teamMember._id)}>Remove</a>
             </tr>
         )
     }
 }
 
-class TeamMembers extends Component {
+DataRow.contextTypes = {
+    removeResource: PropTypes.func
+};
+
+export default class TeamMembers extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            team: {},
+            team: props.team,
             teamMembers: []
         };
 
+        this.removeResource = this.removeResource.bind(this);
     }
 
     componentWillReceiveProps(nextProps) {
@@ -40,34 +43,51 @@ class TeamMembers extends Component {
         });
     }
 
+    getChildContext() {
+        return {
+            removeResource: this.removeResource
+        }
+    }
+
+    removeResource(teamMemberId) {
+        let url = '/team/' + this.state.team.teamName + '/remove/' + teamMemberId;
+        fetch(url)
+            .then(this.setState({
+                teamMembers: this.state.teamMembers.filter(resource => {
+                    return resource._id !== teamMemberId
+                })
+            }))
+    }
+
     render() {
-        let dataRows = this.state.teamMembers.map(function (teamMember) {
-            return <DataRow key={teamMember._id} dataItem={teamMember}/>
+        let dataRows = this.state.teamMembers.map(teamMember => {
+            return <DataRow key={teamMember._id} teamMember={teamMember} teamName={this.state.team.teamName}/>
         });
 
         return (
             <Panel header="Team Members">
-                <Table striped hover responsive>
-                    <thead>
-                    <tr>
-                        <th/>
-                        <th>Name</th>
-                        <th>Role</th>
-                        <th>Action</th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    <tr>
+                {this.state.teamMembers.length ?
+                    <Table striped responsive>
+                        <thead>
+                        <tr>
+                            <th/>
+                            <th>Name</th>
+                            <th>Role</th>
+                            <th/>
+                        </tr>
+                        </thead>
+                        <tbody>
                         {dataRows}
-                    </tr>
-                    </tbody>
-                </Table>
+                        </tbody>
+                    </Table>
+                    : <Alert bsStyle="danger" className="text-center">No team members</Alert>
+                }
+                <br/>
                 <Row>
                     <Col sm={12}>
-                        <Button bsStyle="success" block
-                                href={'/team/' + this.state.team.teamName + '/addTeamMember'}>
-                            Add
-                        </Button>
+                        <LinkContainer to={'/team/' + this.state.team.teamName + '/addTeamResource'}>
+                            <Button bsStyle="success" block><Glyphicon glyph="plus"/></Button>
+                        </LinkContainer>
                     </Col>
                 </Row>
             </Panel>
@@ -75,4 +95,6 @@ class TeamMembers extends Component {
     }
 }
 
-module.exports = TeamMembers;
+TeamMembers.childContextTypes = {
+    removeResource: PropTypes.func
+};
