@@ -3,43 +3,29 @@
  */
 
 import React, {Component} from 'react';
+import {loadDocument} from '../../api';
 import AddTeamResourceSearchResults from '../../components/Resource/AddTeamResourceSearchResults';
 
 export default class AddTeamResourceSearchResultsPage extends Component {
 
-    constructor() {
-        super();
+    constructor(props) {
+        super(props);
         this.state = {
             results: []
         };
 
-        this.performSearch = this.performSearch.bind(this);
-    }
-
-    componentWillMount() {
-        this.performSearch();
-    }
-
-    performSearch() {
-        let url = 'http://localhost:3001/searchResources?resourceName=' + this.props.location.query.resourceName
+        this.loadDocument = loadDocument.bind(this);
+        this.loadDocument('/searchResources?resourceName=' + this.props.location.query.resourceName
             + '&employeeId=' + this.props.location.query.employeeId
             + '&location=' + this.props.location.query.location
             + '&email=' + this.props.location.query.email
-            + '&role=' + this.props.location.query.role;
-        fetch(url)
-            .then(response => response.json())
-            .then((result) => {
-                this.setState({
-                    results: result.results.results
-                });
-            })
-            .catch(err => console.log(err));
+            + '&role=' + this.props.location.query.role);
     }
 
     render() {
         return (
             <div>
-                <AddTeamResourceSearchResults searchResults={this.state.results}/>
+                <AddTeamResourceSearchResults teamName={this.props.params.teamName} searchResults={this.state.results}/>
             </div>
         )
     }
